# src/main/resources/api-definition.yaml
openapi: 3.0.1
info:
    title: Hostsharing hsadmin-ng API
    version: v0
servers:
    - url: http://localhost:8080
      description: Local development default URL.
paths:
...
    /api/rbac-users:
        get:
            tags:
                - rbacusers
            description: List accessible RBAC users with optional filter by name.
            operationId: listUsers
            parameters:
                - $ref: '#/components/parameters/currentUser'
                - $ref: '#/components/parameters/assumedRoles'
                - name: name
                  in: query
                  required: false
                  schema:
                      type: string
            responses:
                "200":
                    description: OK
                    content:
                        'application/json':
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/RbacUser'
                "401":
                    description: if the 'current-user' cannot be identified
                    content:
                        'application/json':
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/RbacUser'
                "403":
                    description: if the 'current-user' is not allowed to assume any of the roles
                        from 'assumed-roles'
                    content:
                        'application/json':
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/RbacUser'
    /api/rbac-users/{userName}/permissions:
        get:
            tags:
                - rbacusers
            description: 'List all visible permissions granted to the given user; reduced '
            operationId: listUserPermissions
            parameters:
                - $ref: '#/components/parameters/currentUser'
                - $ref: '#/components/parameters/assumedRoles'
                - name: userName
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: OK
                    content:
                        'application/json':
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/RbacUserPermission'

                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
...
components:
    parameters:
        currentUser:
            name: current-user
            in: header
            required: true
            schema:
                type: string
            description: Identifying name of the currently logged in user.
        assumedRoles:
            name: assumed-roles
            in: header
            required: false
            schema:
                type: string
            description: Semicolon-separated list of roles to assume. The current user needs to have the right to assume these roles.
    responses:
...
        Unauthorized:
            description: The current user is unknown or not authorized.
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
        Forbidden:
            description: The current user or none of the assumed or roles is granted access to the .
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
    schemas:
...
        RbacUser:
            type: object
            properties:
                uuid:
                    type: string
                    format: uuid
                name:
                    type: string
        RbacUserPermission:
            type: object
            properties:
                objectUuid:
                    type: string
                    format: uuid
                objectTable:
                    type: string
                objectIdName:
                    type: string
                roleName:
                    type: string
                roleUuid:
                    type: string
                    format: uuid
                permissionUuid:
                    type: string
                    format: uuid
                op:
                    type: string
...
        Error:
            type: object
            properties:
                code:
                    type: string
                message:
                    type: string
            required:
                - code
                - message